Add a top level examples folder.
