--- docs/content/api_core/api_Examples.rst
@@ -1,18 +1,0 @@
-.. _api_Examples:
-
-Examples
-********
-
-.. toctree::
-    :maxdepth: 1
-    :glob:
-
-    ../examples/*
-
-
-External Notebooks
-==================
-
-* `Example 1: Direct Current <http://www.seogi.me/s/notebooks/DCEx.html>`_
-* `Example 2: Seismic-Acoustic <http://www.seogi.me/s/notebooks/SeismicEx.html>`_
-
